+++ src/libc/small.c
@@ -2,14 +2,14 @@
 // Symbols expected by gcc/clang in freestanding mode.
 // Put them in separate sections to allow linker to discard unused ones.
 hc_SECTION(".text.memset")
-void *memset(void *dest, int32_t c, uint64_t n) {
+void *memset(void *dest, int32_t c, size_t n) {
     char *d = dest;
     for (; n != 0; --n) *d++ = (char)c;
     return dest;
 }
 
 hc_SECTION(".text.memmove")
-void *memmove(void *dest, const void *src, uint64_t n) {
+void *memmove(void *dest, const void *src, size_t n) {
     char *d = dest;
     const char *s = src;
     if (d < s) {
@@ -24,7 +24,7 @@ void *memmove(void *dest, const void *src, uint64_t n) {
 }
 
 hc_SECTION(".text.memcpy")
-void *memcpy(void *restrict dest, const void *restrict src, uint64_t n) {
+void *memcpy(void *restrict dest, const void *restrict src, size_t n) {
     char *d = dest;
     const char *s = src;
     for (; n != 0; --n) *d++ = *s++;
@@ -32,7 +32,7 @@ void *memcpy(void *restrict dest, const void *restrict src, uint64_t n) {
 }
 
 hc_SECTION(".text.memcmp")
-int32_t memcmp(const void *left, const void *right, uint64_t n) {
+int32_t memcmp(const void *left, const void *right, size_t n) {
     const char *l = left;
     const char *r = right;
     for (;;) {
